requirements.txt
boto3==1.20.24
pandas==1.2.5
psycopg2==2.9.2
python-dotenv==0.20.0
SQLAlchemy==1.4.19
